--- point-api/point-application/src/main/java/com/flab/point/application/TestManager.java
@@ -1,12 +1,0 @@
-package com.flab.point.application;
-
-import com.flab.point.domain.Point;
-
-public class TestManager {
-
-    Point point = new Point();
-
-    public void test() {
-
-    }
-}
