+++ point-api/point-domain/src/main/java/com/flab/point/domain/Point.java
@@ -3,32 +3,32 @@ package com.flab.point.domain;
 import java.time.LocalDateTime;
 
 /*
+ * 사용자별 총 포인트
+ *
  * 포인트 생성 시점
- * 1. 상품 주문 시 1% 적립 포인트 2. 이벤트 포인트 3. 충전
+ * 1. 상품 주문 시 적립 포인트 2. 이벤트 제공 포인트 3. 충전
+ *
+ * 포인트 차감 시점
+ * 1. 포인트 사용 2. 만료 기간 이후 차감
  */
 
 public class Point {
 
     private Long id;
     private Long userId;
-    private PointCategory pointCategory;
     private Long amount;
     private LocalDateTime registerAt;
-
-    // TODO 추후 주문 상품별, 이벤트별 만료 기간 설정. 현재는 적립시 적립 일자 + 한달로 설정
     private LocalDateTime expireAt;
 
 
-    public Point() {
+    protected Point() {
     }
 
     public Point(
             Long userId,
-            PointCategory pointCategory,
             Long amount
     ) {
         this.userId = userId;
-        this.pointCategory = pointCategory;
         this.amount = amount;
         this.registerAt = LocalDateTime.now();
         this.expireAt = setExpireDate();
@@ -36,30 +36,29 @@ public class Point {
 
     // 추가 (적립, 충전)
     public void add(
-            PointCategory pointCategory,
-            Long amount
+            Long addedAmount,
+            PointCategory pointCategory
     ) {
-        this.pointCategory = pointCategory;
-        this.amount += amount;
-        PointTransaction.add(this, amount);
+        this.amount += addedAmount;
+        PointTransaction.add(this, amount, pointCategory);
     }
 
-    // 차감 (사용 시 차감)
+    // 포인트 사용
     public void remove(
-            Long amount
+            Long removedAmount
     ) {
-        Long removedAmount = PointTransaction.remove(this, amount);
+        //Long removedAmount = PointTransaction.remove(this, amount);
         this.amount -= removedAmount;
 
         if (this.amount < 0) {
-            throw new RuntimeException("가용 포인트를 넘어선 접근입니다."); //
+            throw new RuntimeException("가용 포인트를 넘어선 접근입니다."); // TODO exception 처리
         }
     }
 
+    // TODO 추후 주문 상품별, 이벤트별 만료 기간 설정. 현재는 적립시 적립 일자 + 한달로 설정
     private LocalDateTime setExpireDate() {
         return LocalDateTime.now().plusMonths(1);
     }
-
 
     public Long getId() {
         return id;
@@ -67,10 +66,6 @@ public class Point {
 
     public Long getUserId() {
         return userId;
-    }
-
-    public PointCategory getPointCategory() {
-        return pointCategory;
     }
 
     public Long getAmount() {
@@ -81,15 +76,4 @@ public class Point {
         return expireAt;
     }
 
-
-    public enum PointCategory {
-        ORDER("주문 포인트"),
-        EVENT("이벤트 제공 포인트"),
-        CHARGE("사용자 충전 포인트");
-        private final String description;
-
-        PointCategory(String description) {
-            this.description = description;
-        }
-    }
 }
