--- point-api/point-infrastructure/src/main/java/com/flab/point/infrastructure/TestInfra.java
@@ -1,8 +1,0 @@
-package com.flab.point.infrastructure;
-
-import com.flab.point.domain.Point;
-
-public class TestInfra {
-
-    Point point = new Point();
-}
